+++ pybamm/parameters/parameter_sets.py
@@ -21,6 +21,17 @@ Marquis2019 = {
     "citation": "marquis2019asymptotic",
 }
 
+NCA_Kim2011 = {
+    "chemistry": "lithium-ion",
+    "cell": "Kim2011",
+    "anode": "graphite_Kim2011",
+    "separator": "separator_Kim2011",
+    "cathode": "nca_Kim2011",
+    "electrolyte": "lipf6_Kim2011",
+    "experiment": "1C_discharge_from_full_Kim2011",
+    "citation": "kim2011multi",
+}
+
 #
 # Lead-acid
 #