+++ tests/integration/test_models/test_full_battery_models/test_lead_acid/test_full.py
@@ -13,7 +13,9 @@ class TestLeadAcidFull(unittest.TestCase):
         options = {"thermal": "isothermal"}
         model = pybamm.lead_acid.Full(options)
         modeltest = tests.StandardModelTest(model)
-        modeltest.test_all(t_eval=np.linspace(0, 0.6), solver=pybamm.CasadiSolver())
+        modeltest.test_all(
+            t_eval=np.linspace(0, 3600 * 17), solver=pybamm.CasadiSolver()
+        )
 
     def test_basic_processing_with_convection(self):
         options = {"thermal": "isothermal", "convection": True}
@@ -21,7 +23,7 @@ class TestLeadAcidFull(unittest.TestCase):
         var = pybamm.standard_spatial_vars
         var_pts = {var.x_n: 10, var.x_s: 10, var.x_p: 10}
         modeltest = tests.StandardModelTest(model, var_pts=var_pts)
-        modeltest.test_all(t_eval=np.linspace(0, 0.6))
+        modeltest.test_all(t_eval=np.linspace(0, 3600 * 10))
 
     def test_optimisations(self):
         options = {"thermal": "isothermal"}